chore: cleanup route names
