+++ src/router/route-names.ts
@@ -1,0 +1,10 @@
+export const RouteNames = {
+  Home: "Home",
+  Login: "Login",
+  PrintersView: "PrintersView",
+  Settings: "Settings",
+  PrintStatistics: "Print Statistics",
+  About: "About",
+  LackingPermission: "LackingPermission",
+  NotFound: "NotFound",
+};
